Migration tool improvements.
Care is taken to let the progress bar run all the way, in the past failures
caused it to short.

Since we'll mostly be doing xhtml migrations, this adds a utils module with
a factory method that creates migrations for xhtml modules.

If anything fails, the migration continues, but the progress bar turns red
to visually indicate that something was wrong.