Add login modal options to hide details, callback
When invoking the logInModal, you can now request that the anonymous option
and the extra info be hidden, for when you require a login for a specific
change. Use case in mind: Forking while not logged in.

The logInModal method now returns a promise that is resolved upon login,
or rejected otherwise, so that you can hang a follow-up onto a login.